LayoutTests/imported/w3c/web-platform-tests/resize-observer/observe-expected.txt

CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.
CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.

PASS ResizeObserver implemented
PASS guard
PASS test0: simple observation
PASS test1: multiple observation on same element trigger only one
PASS test2: throw exception when observing non-element
PASS test3: disconnect stops all notifications
PASS test4: unobserve target stops notifications, unobserve non-observed does nothing
PASS test5: observe img
PASS test6: iframe notifications
PASS test7: callback.this
FAIL test8: simple content-box observation undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test9: simple content-box observation but keep border-box size unchanged undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test10: simple border-box observation undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test11: simple observation with vertical writing mode undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test12: no observation is fired after the change of writing mode when box's specified size comes from logical size properties. undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test13: an observation is fired after the change of writing mode when box's specified size comes from physical size properties. undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test14: observe the same target but using a different box should override the previous one undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test15: an observation is fired with box dimensions 0 when element's display property is set to inline undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
PASS test16: observations do not fire for non-replaced inline elements
FAIL test17: Box sizing snd Resize Observer notifications undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
FAIL test18: an observation is fired when device-pixel-content-box is being observed undefined is not an object (evaluating 'entries[0].contentBoxSize[0]')
